+++ Tests/LibWeb/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(TEST_SOURCES
     TestCSSIDSpeed.cpp
     TestCSSPixels.cpp
+    TestFetchURL.cpp
     TestHTMLTokenizer.cpp
     TestMicrosyntax.cpp
     TestMimeSniff.cpp
@@ -10,5 +11,7 @@ set(TEST_SOURCES
 foreach(source IN LISTS TEST_SOURCES)
     serenity_test("${source}" LibWeb LIBS LibWeb)
 endforeach()
+
+target_link_libraries(TestFetchURL PRIVATE LibURL)
 
 install(FILES tokenizer-test.html DESTINATION usr/Tests/LibWeb)
